chore: 글 조회 response 변경

+++ src/main/java/com/example/choose_one/model/ViewResponse.java
@@ -16,6 +16,6 @@ public class ViewResponse {
     private String title;
     private String contentA;
     private String contentB;
-    private Integer countA;
-    private Integer countB;
+    private Long countA;
+    private Long countB;
 }
